requirements.txt
torch~=1.10.0
numpy~=1.22.3
scikit-learn~=1.0.2
torchvision~=0.11.1
matplotlib~=3.5.1
Pillow~=9.1.0
einops~=0.4.1
requests~=2.27.1
PyYAML~=6.0
pandas~=1.4.2
seaborn~=0.11.2
wandb~=0.12.12
tqdm~=4.64.0
scipy~=1.8.0
h5py~=3.6.0
opencv-python~=4.6.0.66